Add GradientColorPicker component and enhance ColorPicker modal with gradient selection
