+++ src/components/ui/Gradient.css
@@ -1,0 +1,6 @@
+.gradient-canvas {
+  border-radius: 24px;
+  cursor: crosshair;
+  box-shadow: 0 2px 8px #0002;
+  overflow: hidden;
+}
